(FACT-570) Remove windows-pr dependency
Previously, we used the windows-pr gem to detect which windows version we
were running on, and then either called into the win32-security gem for
windows versions < 6.0 (2003), or used windows-pr gem to determine if
we were running with elevated privileges.

However, the `Win32::Security.elevated_security?` method now correctly
handles Vista and up, so we can just call that directly, avoiding
windows-pr, and in the process eliminating win32-api.