(maint) Update the behavior of the domain/fqdn facts to match Ruby facter.
Recently Ruby facter was changed so that the fqdn fact would resolve to the hostname
if the domain name was empty.

Our implementation differs by actually resolving the hostname with the default name
server.  When the resolution failed with a "no name" error, we would emit a warning
and not resolve either the domain or fqdn facts.

This commit changes the warning to imply that we failed to resolve the hostname and
that will likely indicate that the fqdn fact is not externally resolvable. We then
default the fqdn fact to the hostname and resolve the domain based on the value
of fqdn.