(maint) Update packaging
We're also building with pl-cmake rather than cmake, and linking against
the correct pl-gcc libraries. There were also some missing dependencies
that were missing from the original packaging update. We also need to
update the rpath for 64bit, so we link against the right libraries

Additionally, we've update the toolchain provided by cmake, so we don't
need to specify anything in order to take advantage of the build tools
we've created. The toolchain tells cfacter where all the libraries are
that it needs, and what to use to build. IT'S AWESOME