(maint) Move `lib` subdirectory add to top-level cmake
Previously, the `lib` dir was added from the exe CMake. This is not
idiomatic (add_subdirectory should be used only for true
subdirectories), but also gives a structural impression of "We are
building an executable that contains a library". Moving the `lib`
subdir add to the top-level gives the more accurate "We are building a
library and an executable (which happens to use our library)"