(maint) Add a can_resolve member to external fact resolvers.
Currently the resolve member of external fact resolvers returns false if
the resolver isn't responsible for files of the given type.

This logic belongs in a "can_resolve" member that can be used to check
if the resolver can resolve facts in the given file.

This allows the fact collection to be able to map an external facts file
to the responsible resolver.