(FACT-569) refactor windows acceptance nodesets
This commit refactors Facter Windows acceptance testing configurations
for Beaker. Prior to this commit, each of these files contained multiple
Windows nodes and Ruby architectures. This commit splits them into
separate files so that:
- Facter tests can run in parallel
- Easily see which platforms and Ruby arch's in the matrix passed/failed

This commit also reduces our testing to Windows R2 releases only, except
in the case of Windows 2003; we'll continue to test non-R2 for that
version.