(maint) Fix always executing system uptime command.
The recent addition of the structured uptime fact caused the uptime
command to always be run.  This is because resolve_system_uptime always
called executable_uptime.  Now that these facts are no longer dependent
on the seconds fact and are now dependent on the structured fact, there
should be a virtual member that simply returns the uptime in seconds
without adding any facts.