(FACT-627) Print `false` when outputting a single fact value
Before this commit the plaintext formatter did a naive `if value` which
is false when `value` is either nil or false. While we don't want to
print nil facts, facts that are false are valid but the previous
behavior would treat them as undefined as well. This commit ensures that
the fact value is not nil, and if the fact is defined then it is
printed.

Thanks to RI Pienaar and Robin Bowes for finding and reporting this
issue.