(maint) Fix string value output quoting.
When outputting string values, we should default to having them quoted
instead of always quoting values in debug output.  This makes it easier
for users to differentiate between values that aren't strings and values
that are.  For example, a boolean value of false should be printed
without quotes, but a string value of "false" should be printed with
quotes.

Refactoring the writing of values to a stream to control
whether or not string values should be quoted.