(maint) Don't assume processor facts are defined
If Facter cannot resolve the processors fact (eg Facter.value is stubbed
or the processors fact isn't implemented on a platform) we cannot always
assume that we have a hash. This commit changes the enumerated processor
facts to retrieve the processors fact and ensure that it's defined before
invoking methods on it.