(maint) Lazily evaluate and memoize processor implementation
Without this commit the processors fact will try to determine the OS
specific implementation when the fact is loaded, which triggers a lot of
side effects at load time. This commit converts the fact to lazily load
and memoize the object so that loading is deferred as much as possible.