(maint) Change Ruby API to be a singleton.
Originally the Ruby API was designed to be reloadable.
However, dlclose may not unload the Ruby library on some platforms.
As a result, the Ruby library remains resident and in a state that makes
it impossible to properly reinitialize.

Therefore, the Ruby API should be changed to a singleton that can only
be constructed and destructed once.