(FACT-613) Convert 'models' key of processors structured fact to array
This commit changes the 'models' key within the processors structured
fact into an array. Previously, the key was a hash which was used
to support the legacy dynamic 'processorX' flat facts. With this
change, the flat fact itself handles its naming based on the index
number of a processor in the array.