(CFACT-71) Fix Ruby VM initialization.
Previously, intialization of the Ruby VM was occuring at a lower stack
frame than where we were loading code into the VM.  This caused a gap on
the stack where the Ruby GC would not search for VALUEs to mark,
resulting in random crashes when Ruby would GC objects still in use.

The fix is to properly initialize Ruby from main, thereby ensuring that
it gets initially configured with the highest possible stack location.
This will ensure the GC searches the entire stack for VALUEs to mark.