(CFACT-44) CFacter gem should always return a Fact object.
Unlike Facter, CFacter doesn't make a distinction between facts that
loaded but didn't resolve (value is nil) and facts that didn't load.

Facter#[] returns nil for facts that didn't load and a Fact object for
facts that loaded but may not have resolved (value is nil).

Therefore, CFacter#[] needs to always return a fact object to keep this
expected behavior.