(CFACT-44) Replace cfacter gem with libfacter Ruby API.
The cfacter gem implementation is unnecessary now that we implement the
Ruby API in libfacter.  Therefore, remove it in favor of directly
calling into libfacter from Ruby.

To implement the gem's behavior, the C++ Facter module implementation
now implements the following methods:

* Facter.flush (and implemented on_flush in resolutions)
* Facter.list
* Facter.to_hash
* Facter.each
* Facter.clear
* Facter.reset
* Facter.loadfacts
* Facter.search
* Facter.search_path
* Facter.search_external
* Facter.search_external_path

Additionally, the logic for loading Ruby facts has moved out of the
collection (where it didn't belong) and moved into the C++ Facter
module.

We now follow the old Facter's loading design of first trying
to load a fact by file name and then falling back to loading all facts
if the fact can't be found.