(CFACT-44) Allow Ruby classes to be instantiated from Ruby.
Currently the Ruby classes implemented in libfacter can only be
instantiated from C++. If one of the classes is instantiated from Ruby,
we fail to associate the Ruby instance with the underlying C++ type.

To fix this, define an allocation, mark, and free callback for each type
we support allocating from Ruby.  The allocate function is responsible
for creating the underlying C++ type when the Ruby type is allocated.
The mark function is responsible for marking referenced VALUEs during
the mark phase of GC.  The free function is responsible for freeing the
underlying C++ type when the GC deallocates the Ruby object.