(CPR-5) Move dmidecode from dependency to recommends
Previously facter had a hard dependency on dmidecode. This made the
debian package uninstallable on ARM based devices such as raspberry pi.
This commit moves dmidecode from dependency to recommends. In nearly all
cases dmidecode will still be pulled in, unless it can't be (as in the
ARM case). This also moves the package from all (meaning any arch it can
be built for) to any (meaning noarch).

Note this is a packaging change only.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>