(CFACT-81) Fix LSB major/minor release resolutions in Ubuntu
Prior to this commit, the LSB release facts were not correctly
parsing Ubuntu's release string to collect its major and minor
versions. This commit updates the LSB facts such that a release
string of '14.04.1' will return a major release of '14.04' and
a minor release of '04', if a minor version is specified.