(CFACT 70) Add kernel facts for solaris
Before this patch, the Solaris operating system did not produce kernel facts.
Unfortunately the posix kernel facts could not be used as it relied on the uname()
system call, which according to posix, returns zero on success. However, on solaris,
it returns it returns any positive value on success[1]

This patch adds the solaris kernel facts which checks the return value of uname
system call correctly for solaris.

[1] http://www.opensolarisforum.org/man/man2/uname.html