(FACT-683) Use XPG4 variant of `id` in Solaris to determine GID
Prior to this commit, the GID fact would attempt to invoke the `/usr/bin/id`
variant of the `id` command in Solaris, which does not include the necessary flags to
determine the user GID in Solaris 10.

This commit updates the fact to instead use the more advanced `/usr/xpg4/bin/id`
for Solaris instead.