(FACT-692) Fix issue with Solaris resolution of operatingsystemmajrelease
This commit corrects the naming of the method responsible for
determining the major release of SunOS systems, allowing Solaris 10
systems to correctly resolve the operatingsystemmajrelease fact.

With the addition of the structured operatingsystem fact, a new
class structure was created which includes a base class with
several platform specific children. The method within these
classes which is responsible for determining the operatingsystem
major release is named 'get_operatingsystemmajorrelease'. The
method was, however, misnamed in the SunOS class as
'get_operatingsystemmajrelease', which is never called, causing
the fact to fall back to the base class implementation.

This commit fixes this regression which was introduced in Facter 2.2.0
at commit 793b258.