(CFACT 70) Add processor facts for solaris
This patch adds solaris specific processor facts. For this it utilizes the
kstat linked datastructure, which is added in by kstat library. Solaris
has the concept of three levels of processing units: physical CPUS, multiple
cores per CPU and, multiple hardware threads per core. This is reflected in
the provided facts.