(CFACT 70) Add filesystem facts for solaris
Before this patch, solaris did not provide the filesystem facts. This patch
adds support for filesystem facts for solaris. Unlike linux, there is no direct
means to find supported filesystems, and hence we use the bsd approach, and
list only the currently in use filesystems. We also leave the partitions
support unimplemented for now.