(maint) Fix Linux partition fact attribute names.
During the Boost function refactoring, tag_value was used for the
attribtue instead of tag_name.  This resulted in the partitions fact
having correct values, but incorrect keys (the keys were the same as the
value).

The fix is to use the appropriate name coming back from blkid_tag_next.