(CFACT-107) Sync Solaris operatingsystem facts with structured OS fact
As part of the structuring of the operating system facts into the
'os' structured fact, the implementations of the operating system
resolvers for POSIX, Linux and OSX were updated such that the
structured fact itself would interrogate the system and make its values
available for the older flat facts to collect their values from.

The Solaris operating system flat facts were implemented at the
same time that this work was happening, and thus did not get the
same treatment with regards to structuring. This commit updates
the Solaris operating system resolver so that it is in line with
other class implementations, and so that the structured OS fact
resolves correctly in Solaris.