(CFACT-121) Implement find_by_pattern and minor fixes
Implemented find_by_pattern, which searches loaded DLLs for one that matches the
specified pattern and returns a handle to it in the form of a dynamic_library.
Fixed const and & correctness of helper functions.
Changed Ruby library version comparisons to use int instead of string.
Check that a sub_match is valid before using lexical_cast in regex.hpp, so optional
matches don't throw an exception.