(CFACT-95) Update Solaris operatingsystemrelease value to match Facter 2.2
In Facter 2.2, we used different formatting for the operatingsystemrelease
fact depending on which version of Solaris the fact was being resolved
in. The initial implementation of the Solaris operatingsystemrelease
fact in native Facter changed this so that the OSrelease would always
resolve as <major>_u<minor>.

In order to maintain backwards compatibility and be consistent with
previous versions of Facter, this commit updates the
operatingsystemrelease resolution in Solaris so that in Solaris 10 the
string will read 'X_uY', while in Solaris 11.0 and newer it
will read '11.Y'.