(CFACT-124) Refactor operating system resolver.
Refactoring the operating system resolver into a base that is
responsible for adding the facts and child types that are responsible
for collecting data.  This allows it to be shared with Windows.

This refactoring also merges the LSB resolver in with the operating
system resolver.

Breaking change: some of the names on the os structured fact have
diverged from Facter 2.2 to clean up the less than ideal naming. We need
to investigate if this will be impactful and how to mitigate.