(CFACT-128) Switch to *nix-agreed implementation of timezone reporting
Two things to fix here:

1) The daylight variable seems to be non-standard. At least it's not
mentioned here:
http://pubs.opengroup.org/onlinepubs/7908799/xsh/time.h.html
which states "The value of tm_isdst is positive if Daylight Saving Time
is in effect, 0 if Daylight Saving Time is not in effect, and negative
if the information is not available." (I speculate the use of a daylight
extern'd by the C library is an early implementation.)

2) FreeBSD doesn't support tzname: https://wiki.freebsd.org/FreeBSD_and_Standards

Next I turned to gettimeofday, but no love from FreeBSD there either:
http://www.freebsd.org/cgi/man.cgi?query=gettimeofday&sektion=2

Next I looked at boost, which allows trick features like custom timezones but not
(apparently, I may have missed it) simple reporting of current timezone.

Next up was std::time_put, but that's not implemented in gcc 4.8:
http://stackoverflow.com/a/14142342/1004272

So, crossing fingers, but good ol' strftime may do the trick. Yippee.