(CFACT-91) Add test asserting Powershell architecture matches compilation
We want to use 64-bit Powershell on 64-bit Windows. The Powershell used is determined by the architecture of the process
executing it, by hidden Windows redirection. This test confirms that 64-bit Powershell is used for a 64-bit compiled
facter, and 32-bit Powershell is used otherwise. This should cover most desired cases of providing 64-bit Powershell
to users.