(CFACT-91) Test for and filter trailing separators in PATH
Trailing separators in PATH would be concatenated with a relative path
in filesystem::path constructors, which would give a valid relative
path. This led to environment-dependent behavior, as relative paths are
not normally accepted by 'which'. Fix this by stripping trailing
separators and testing that we don't have empty path segments.