(FACT-699) Create system32 fact to resolve native system32 directory
Adds a `system32` fact which resolves the "real" system32 directory for both
32 and 64-bit ruby versions. This way you can write a manifest that simply
does:

    exec { "$::system32\\cmd.exe ...": }

and it will always execute 64-bit cmd.exe on 64-bit operating systems,
regardless of the ruby architecture facter happens to be running on.

Note the fact is implemented by stat'ing the filesystem, because there is no
CSIDL constant that resolves to sysnative. For example, Dir::SYSTEM resolves to
'C:\Windows\system32' and Dir::SYSTEMX86 resolves to 'C:\Windows\SysWOW64',
but there isn't a Dir::SYSNATIVE.