(maint) Remove unused cycle detection in fact resolvers.
Because of the way we need to force a resolver to resolve when a fact it
is responsible is added to the collection (i.e. from a custom or
external fact), we remove a resolver from the collection prior to doing
the resolve.

As a result, the old cycle detection isn't used because the resolver
isn't in the collection during resolution.  Instead of throwing a cycle
detection, we break the cycle and return nullptr for the requested fact.