(CFACT-147) Update dependencies
We want to make sure we have the correct dependencies at the correct
times. Since we are shifting to building cfacter with static libraries,
we need to ensure that the packages used at build time actually contain
those static libs. The main reason to statically link against some
libraries is that some are not publically available. In order to get
these packages into the hands of our users, we need to remove any
runtime dependencies that are limitied to them. This change allows us to
do that.

This commit also adds in some of the optional cfacter dependencies,
enabling their functionality during runtime.