(CFACT-147) Build cfacter explicitely with the static libs of boost and yaml-cpp
This updates cmake to specifically search for and link against the
static libraries of boost and yaml-cpp. This is because we have both
versions available on the system. Since the linker by default will
prefer the shared libs, we need to be explicit about our desire for the
static libs.