(maint) Fix detection of xenu hypervisor.
The code was checking for the existence of only a file instead of a
file or directory like we have under Ruby Facter.

This caused us to fail to detect the xenu hypervisor (noticed on EC2,
which uses Xen).  The fix is to change the detection to just an "exists"
call to make the Ruby behavior.