(CFACT-149) build for el 6
openssl-libs doesn't exist on el6, so we need to pull in openssl
instead to provide those libraries.

We're also updating dep reqs to make sure we're pulling in the most
recent versions of dependencies, that have the static pacakge available.

We're also adding in libcurl at build time so that we can enable that
functionality.