(CFACT-18) Implement EC2 facts.
Implementing the ec2_metadata and ec2_userdata facts in libfacter.

These facts will be resolved when under a kvm/xen hypvervisor and the
BIOS version has "amazon" in it.

The ec2_metadata and ec2_userdata facts match exactly to what is output
in Ruby facter.

As the "flattened" facts from ec2_metadata were marked deprecated in
Ruby facter, we've dropped support for them.