(FACT-705) Improve documentation of structured facts
Prior to this commit, the Facter core documentation did not
specify which facts were structured. In addition, the documentation
for several structured facts did not adequately describe the contents
of the fact.

This commit improves the header comments for several structured facts,
which are used to auto-generate the Facter core docs. Formatting was
used in accordance with the tool used to generate our core docs, which
can be found here: https://github.com/holguinj/extracter.