(CFACT-155) Fix POSIX execution to redirect to a writable /dev/null.
When not redirecting stderr to stdout, libfacter redirects stderr to
/dev/null on POSIX platforms.  However, the open call was passed the
read-only flag, making the stderr descriptor unwritable.

If a child process were to check if write failed on stderr (lsb_release
does, for example), then the process would fail when it would otherwise
succeed.

The fix is to allow the child process to write to the descriptor.