(maint) Specifically link against rt library
Address failure:
`../libfacter.so.0.2.0: undefined reference to `clock_gettime'
collect2: error: ld returned 1 exit status`

It looks like we are incorrectly linking against the rt library. In
newer el versions, it is likely that this gets pulled correctly, but for
older versions of el, this doesn't seem to be the case.