(CFACT-156) Build CFacter for el5
Unfortunately, el5 doesn't have any of the "optional" buildtime
requirements. It does have curl and openssl development packages
available, however, these are both of insufficient versions (curl 7.15.5
and openssl 0.9.8e as opposed to curl >= 7.18.0 and openssl >= 1.0.1g).
In light of this, we will not build with these dependencies for el5.