(maint) Allow for easier building on AIX
This is something to help us in the future. Eventually, we will use this
spec file to build packages for AIX. Since AIX is very particular about
the syntax is supports, these are some very small changes to comply with
that. Nested conditionals are a no no. Obviously, this is only the tip
of the iceberg when it comes to changes that will be needed for AIX, but
it's nice to have one less thing to think about when the time comes