(FACT-658) Update uptime regexes for Solaris edge cases
In some versions of `uptime` on some versions of Solaris,
output may differ from what we previously expected. Specifically,
the coreutils GNU version does not include a ',' after 'day(s)', and
in some cases uptime may return a '-' character in the minutes field.

This commit updates the relevant regular expressions so that we can
match and parse these uptime strings.