(maint) Define a minimum of 1 job to run
Prior to this commit, on some systems, `nproc` wasn't available on the
system. Since we use this to define the job threading for building with
cmake, having nproc not be defined left us calling `make -j`. This
basically tells make to build with an unlimited number of jobs, eating
up all available system memory, making the build fail with `virtual
memory exhausted: Cannot allocate memory`. This commit changes how we
are defining the number of jobs to thread at once. We want to use
`nproc` when we can, but when we cannot, we just set it at `1`. Ideally
we build with `nproc+1`, so this ensures that we are not eating up all
available memory, and are running jobs at maximum efficiency when
possible.