(maint) Update buildtime requirements
CFacter has a soft requirement on openssl >= 1.0.1g, but it will still
build with other versions of openssl. Since we want to enable as much
functionality in our builds, we want to link against openssl if
possible. Older versions of debian/ubuntu only have openssl in the 0.9.8
range, so when it comes to runtime dependencies, we want to allow the
user to install whatever libssl is available on the system. In this
case, we have a choice between libssl1.0.0 (for openssl1.0.1) or
libssl0.9.8.

Also, it was a little silly that we were using libcurl4-gnutils, rather
than libcurl4-openssl. I believe the original choice was rather
arbitrary, but since we're using openssl here anyway, it makes sence to
use that version of libcurl.