(CFACT-145) Fix mismatch between Ruby and C++ log level enum.
Recently the "none" logging level was added to completely disable
logging by libfacter.

Unfortunately I forgot to update the corresponding definition in the
Ruby gem.  This mismatch caused the "info" level to be the
default during libfacter initialization instead of the expected
"warning" level.  This resulted in unexpected messages to be printed out
by libfacter when run under Puppet.

The fix is to properly synchronize the two enum definitions.