(maint) Update README to link to extensibility document.
Updating the README to link to our extensibility document so that users
know how to extend native facter using custom and external facts.

Removing the C++ usage section from the README; I think that this makes
it seem that you extend native facter in C++, and that's not the case.