(maint) Remove boost nowide from packages
boost-nowide is a package we vendor via submodules. This vendored
package gets build for use when we build cfacter. The libraries it
creates are only needed for buildtime, and not at runtime. In order to
only ship files that are needed, we should pull out all of the files
that boost-nowide installed. This pull request does that.