(FACT-753) Update default external fact directories for unified agent layout
This commit updates the default search directories for external facts.
We now look in /opt/puppetlabs/agent/facts.d for both open source and
PE versions of Facter. When Facter 3 ships, we will default to only
searching in this directory, but to facilitate a smoother transition
Facter 2.4 will look in both old and new external fact search paths.

Note that this change only affects Unix agents, as Windows paths are
staying the same.

This change is part of the unified agent effort.