(CFACT-150) Update default external fact directory for unified agent layout
This commit updates the default search directory for external facts.
We now look in /opt/puppetlabs/agent/facts.d for both open source and
PE versions of Facter. Note that this change only affects Unix agents,
as Windows paths are staying the same.

This change is part of the unified agent effort.