(FACT-184) Facter can't detect long interface names
The current way how to detect interface names is using the "ifconfig"
but this command shortened the long interface names. As Facter reads
"ifconfig" output with a shortened interface names puppet is not
able to work with them anymore. There is a patch that uses "ip link"
instead of "ifconfig" to retrieve interface names because "ip link"
provides the full interface names.