(maint) Ensure we are always running verbose tests
Tests had already been run during the debian build process, but this was
not happening during the rpm build process. This commit makes sure that
we are running tests when building our rpm packages, and has the debian
tests run in verbose mode, so we can get more information on them if
they fail.