(maint) We are not using an epoch to version cfacter
This was copy pasta that was brought in accidently when I first added in
the build artifacts for cfacter. The digit I am removing here is meant
to correspond to the epoch number of the build. This was required for
facter when we previously messed up the versioning. This is not required
for cfacter. Since this error was only present in the changelog, not the
actual body of the spec file, this change should not impact users.