(CFACT-178) Fix segfault in gem cleanup.
When libfacter is initialized via the gem, a new fact collection and
Ruby module are created to provide facts.  We use a "global" object with
an associated data object to track when the Ruby VM shuts down.  When
Ruby exits, we get called to free the data object, allowing us to
cleanup previously allocated resources.

However, this presents a problem because FFI uses a data object to
reference libfacter.  When that data object is freed, libfacter will be
unloaded (this is especially apparent on Windows as the DLL is always
unloaded from memory).  Therefore, if the FFI data object is freed
before our data object, Ruby may call back into a memory page that is
no longer mapped into the process, resulting in a segfault.

The fix is to use a finalizer instead of a data object.  In all
supported versions of Ruby, finalizers are called, at exit, prior to
freeing any data objects.  This behavior guarantees that we can clean up
our resources before FFI unloads us.

This applies to all data objects used in libfacter, including those for
facts and resolutions (aggregate and simple).