(maint) Remove LOG_NAMESPACE macro.
Removing the LOG_NAMESPACE redefinitions from libfacter.  We now use a single
namespace for all of libfacter's output.

This work is needed if we plan on extracting libfacter's logging into a
common library.  The specific namespaces were not adding any value for
libfacter users and only made the code a little more complicated and
brittle in terms of unity builds.