(CFACT-181) Build cfacter for both arches
We need to provide builds of cfacter that support both 64-bit and 32-bit
builds. Our current script is tailored heavily to building for 64-bit.
This updates the script to allow for users to pass in arguments on the
command line to specify what architecture that would like to build for.