(CFACT-200) Fix output of custom fact values with UTF8 characters.
When outputting custom fact values, we were using String#size to get the
size of the strings to write to JSON or to an output stream.

This uses the character size of the string, not the byte size of the
string.  As such, strings were incorrectly truncated when multi-byte
characters were present in the string.

The fix is to use String#bytesize to properly write out all the bytes in
the string.