(CFACT-199) Implement support for setting fact values from environment.
Implementing support for setting fact values from environment variables.

Now users can set variables of the form "FACTER_xyz" and a string fact
will be added.  Environment variable facts are the last source of truth
and can override built-in, external, and custom fact values.

This is supported for backwards compatibility with previous versions of
Facter.